chore: Refactor Reports Endpoints to a Domain.
