+++ src/Reports/Endpoints/AverageDonation.php
@@ -6,12 +6,12 @@
  * @package Give
  */
 
-namespace Give\API\Endpoints\Reports;
+namespace Give\Reports\Endpoints;
 
 use WP_REST_Request;
 use WP_REST_Response;
 
-class AverageDonation extends Endpoint {
+class AverageDonation extends BaseEndpoint {
 
 	protected $payments;
