+++ src/Reports/Endpoints/FormPerformance.php
@@ -6,9 +6,9 @@
  * @package Give
  */
 
-namespace Give\API\Endpoints\Reports;
+namespace Give\Reports\Endpoints;
 
-class FormPerformance extends Endpoint {
+class FormPerformance extends BaseEndpoint {
 
 	protected $payments;
