+++ src/Reports/Endpoints/IncomeBreakdown.php
@@ -6,9 +6,9 @@
  * @package Give
  */
 
-namespace Give\API\Endpoints\Reports;
+namespace Give\Reports\Endpoints;
 
-class IncomeBreakdown extends Endpoint {
+class IncomeBreakdown extends BaseEndpoint {
 
 	protected $payments;
